Dropbox.munki: Don't block if not running from /Applications
Dropbox's auto-updater stores the updated app in the user's Library folder if it can't write to /Applications. That means that we only need to block the installation if the user is running the system-installed Dropbox.app and we can safely update it if the user is running his own updated copy.