Update munkitools6.munki.recipe (#458)
* Update munkitools6.munki.recipe

Munki 6.3, released yesterday, changed the minimum supported version of macOS to 10.13.

https://github.com/munki/munki/releases/tag/v6.3.0

* Update munkitools6.munki.recipe

Updates the Description and includes the variable option I suggested as an alternative to #459.

* Update munkitools6.munki.recipe

Adds optional code signature verification. Thanks to @gregneagle for the suggestion.

Also converts a few random tabs to spaces for consistency.

* Update munkitools6.munki.recipe

Includes requested changes to the `DISABLE_CODE_SIGNATURE_VERIFICATION` variable and true/false.

* Create munkitools6-signed.munki.recipe

Pivots back to separate recipes for the normal and the signed versions of Munki based on discussion in #autopkg on the Mac Admins Slack.

Includes other needed changes/fixes in the existing recipe and updates the Description to reference the new recipe.